[BUGFIX] Special Season episodes
Special episodes outside normal seasons are typically grouped under
season '0'. As such when using an 'or' condition against the value
of the season number or using parent primary key will result in
an attribute error. '0' is considered false in python and then
the parent for episodes is always None.

fixes #38